requirements.txt
asgiref==3.6.0
Django==4.1.6
sqlparse==0.4.3
djangorestframework==3.11.0
requests
redis==4.2.2
pytz
PyMySQL~=1.0.2
uwsgi==2.0.18
yt_dlp
six~=1.16.0
future~=0.18.2
django-crispy-forms~=1.14.0
django-formtools
httplib2~=0.20.4
django-import-export
cryptography==41.0.1
pycryptodome